Feature: 계산 로직을 util 함수로 분리
