+++ src/refactoring/utils/discount.ts
@@ -1,0 +1,14 @@
+import { CartItem } from '../../types.ts';
+
+// 주어진 장바구니 아이템에 대해 적용된 할인율을 계산하는 함수입니다.
+export const getAppliedDiscount = (item: CartItem) => {
+  const { discounts } = item.product;
+  const { quantity } = item;
+  let appliedDiscount = 0;
+  for (const discount of discounts) {
+    if (quantity >= discount.quantity) {
+      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+    }
+  }
+  return appliedDiscount;
+};
